style(frontend): organize components
